+++ src/main/java/net/mysticcloud/spigot/core/utils/punishment/PunishmentUtils.java
@@ -26,7 +26,7 @@ public class PunishmentUtils {
 		ResultSet rs = CoreUtils.sendQuery("SELECT * FROM Punishments;");
 		try {
 			while (rs.next()) {
-				int duration = Integer.parseInt(rs.getString("DURATION"));
+				long duration = Long.parseLong(rs.getString("DURATION"));
 				UUID uid = UUID.fromString(rs.getString("UUID"));
 				PunishmentType type = PunishmentType.valueOf(rs.getString("ACTION"));
 				long date = Long.parseLong(rs.getString("DATE"));
